55535_arbitrary_proto/1_example/requirements.txt
grpcio-tools

ray[serve]
